refactor(setresponse 함수 분리): commit
1. errorpage internal redirection 위함.

@ref https://nginx.org/en/docs/http/ngx_http_core_module.html#error_page

Event 내부에 변수 하나 플래그 두고, setResponse의 일부 함수만
호출하게끔해야할듯.
